% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_p.R
\name{calculate_p}
\alias{calculate_p}
\title{Title}
\usage{
calculate_p(
  G25,
  G_Ea,
  Jmax25,
  Jmax_dS,
  Jmax_Ea,
  Jmax_Hd,
  Km25,
  Km_Ea,
  Temp,
  Vcmax25,
  Vcmax_dS,
  Vcmax_Ea,
  Vcmax_Hd
)
}
\arguments{
\item{G25}{}

\item{G_Ea}{}

\item{Jmax25}{}

\item{Jmax_dS}{}

\item{Jmax_Ea}{}

\item{Jmax_Hd}{}

\item{Km25}{}

\item{Km_Ea}{}

\item{Temp}{}

\item{Vcmax25}{}

\item{Vcmax_dS}{}

\item{Vcmax_Ea}{}

\item{Vcmax_Hd}{}
}
\value{

}
\description{
Title
}
