% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_photo.R
\name{calculate_photo}
\alias{calculate_photo}
\title{Title}
\usage{
calculate_photo(
  data_phys,
  data_env,
  varnames = list(GammaStar = "GammaStar", Jmax = "Jmax", Km = "Km", leaf_respiration =
    "leaf_respiration", root_respiration = "root_respiration", shoot_respiration =
    "shoot_respiration", total_respiration_plant = "total_respiration_plant", Ca = "Ca",
    Qin = "Qin", VPD = "VPD", Temp = "Temp", Vcmax = "Vcmax"),
  alpha = 0.8,
  g0 = 0.0225,
  g1 = 7.7527,
  leaf_area,
  phi,
  time_step
)
}
\arguments{
\item{data_phys}{}

\item{data_env}{}

\item{varnames}{}

\item{alpha}{}

\item{g0}{}

\item{g1}{}

\item{phi}{}

\item{time_step}{}
}
\value{

}
\description{
Title
}
