% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_functions.R
\name{fit_medlyn}
\alias{fit_medlyn}
\title{Title}
\usage{
fit_medlyn(data, varnames = list(k = "k", Temp = "Temp"))
}
\arguments{
\item{data}{}

\item{varnames}{}
}
\value{

}
\description{
Title
}
\examples{

}
