% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_functions.R, R/t_functions.R
\name{fit_new}
\alias{fit_new}
\alias{Arrh_medlyn}
\alias{Arrh_new}
\alias{Arrh_j42}
\alias{Arrh_harley86}
\alias{Arrh_harley92}
\alias{Arrh_farquhar}
\alias{Arrh_topt}
\alias{Q10_funct}
\alias{Arrh_base}
\title{Temperature response functions}
\usage{
fit_new(data, varnames = list(k = "k", Temp = "Temp"))

Arrh_medlyn(k25, delS, Ea, Hd, Temp)

Arrh_new(k25, delS, Ea, Hd, Temp)

Arrh_j42(c, Temp, Ea, delS, Hd)

Arrh_harley86(k25, c, Ea, delS, Hd, Temp)

Arrh_harley92(c, Ea, delS, Hd, Temp)

Arrh_farquhar(c, Ea, delS, Hd, Temp)

Arrh_topt(kopt, Ea, Hd, Topt, Temp)

Q10_funct(k25, Q10, Temp)

Arrh_base(k25, Ea, Temp)
}
\arguments{
\item{k25}{Rate of process at 25 Celsius}

\item{delS}{Entropy parameter in kJ/mol}

\item{Ea}{Activation enegy in kJ/mol}

\item{Hd}{Deactivation energy in kJ/mol}

\item{Temp}{Temperature in Kelvin}

\item{c}{Scaling constant for the Harley et al. 1992 temperature response}

\item{Q10}{Thermal sensitivity of process, representing X-fold change per
10 Celsius change in temperature}
}
\value{
Provides functions for modeling temperature responses of biological
processes. Includes Arrhenius (1915), canonical relativized modified Arrhenius 
(Medlyn et al., 2002 based on Johnson et al. 1942), canonical unrelativized 
modified Arrhenius (Farquhar et al., 1980; Harley et al., 1986, 1992), 
rederived modified Arrhenius (from Johnson et al., 1942), and a Q10 response 
for respiration.

Arrhenius S. 1915. Quantitative laws in biological chemistry. Bell.

Farquhar GD,von Caemmerer S, Berry JA. 1980. A biochemical model of 
photosynthetic CO2 assimilation in leaves of C3 species. Planta 149:78-90.

Harley PC, Tenhunen JD, Lange OL. 1986. Use of an analytical model to study 
limitations on net photosynthesis in Arbutus unedo under field conditions. 
Oecologia 70:393-401.

Harley PC, Thomas RB, Reynolds JF, Strain BR. 1992. Modelling photosynthesis
of cotton grown in elevated CO2. Plant, Cell & Environment 15:271-282.

Johnson FH, Eyring H, WIlliams RW. 1942. The nature of enzyme
inhibitions in bacterial luminescence: sulfanilamide, urethane,
temperature and pressure. J Cell Comp Physiol 20:247-268

Medlyn BE, Dreyer E, Ellsworth D, Forstreuter M, Harley PC,
Kirschbaum MUF, Le Roux X, Montpied P, Strassemeyer J, Walcroft A,
Wang K, Loutstau D. 2002. Temperature response of parameters of a
biochemically based model of photosynthesis. II. A review of
experimental data. Plant Cell Environ 25:1167-1179
}
\description{
Temperature response functions
}
