% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_p.R
\name{calculate_p}
\alias{calculate_p}
\title{Calculating photosynthetic parameters}
\usage{
calculate_p(
  G25,
  G_Ea,
  Jmax25_m,
  Jmax_dS_m,
  Jmax_Ea_m,
  Jmax_Hd_m,
  Jmax25_n,
  Jmax_dS_n,
  Jmax_Ea_n,
  Jmax_Hd_n,
  Km25,
  Km_Ea,
  Temp,
  Vcmax25_m,
  Vcmax_dS_m,
  Vcmax_Ea_m,
  Vcmax_Hd_m,
  Vcmax25_n,
  Vcmax_dS_n,
  Vcmax_Ea_n,
  Vcmax_Hd_n
)
}
\arguments{
\item{G25}{Photorespiratory CO2 compensation point (umol mol-1)}

\item{G_Ea}{Activation energy of G25 (kJ mol-1)}

\item{Jmax25_m}{Maximum electron transport rate at 25 Celsius for the
Medlyn et al. 2002 modified Arrhenius model}

\item{Jmax_dS_m}{Entropy parameter of Jmax (kJ mol-1) for the
Medlyn et al. 2002 modified Arrhenius model}

\item{Jmax_Ea_m}{Activation energy of Jmax (kJ mol-1) for the
Medlyn et al. 2002 modified Arrhenius model}

\item{Jmax_Hd_m}{Deactivation energy of Jmax (kJ mol-1) for the
Medlyn et al. 2002 modified Arrhenius model}

\item{Jmax25_n}{Maximum electron transport rate at 25 Celsius for the
rederived modified Arrhenius model}

\item{Jmax_dS_n}{Entropy parameter of Jmax (kJ mol-1) for the
rederived modified Arrhenius model}

\item{Jmax_Ea_n}{Activation energy of Jmax (kJ mol-1) for the
rederived modified Arrhenius model}

\item{Jmax_Hd_n}{Deactivation energy of Jmax (kJ mol-1) for the
rederived modified Arrhenius model}

\item{Km25}{Michaelis-Menten coefficient for rubisco at 25 Celsius}

\item{Km_Ea}{Activation energy of Km (kJ mol-1)}

\item{Temp}{Temperature in K}

\item{Vcmax25_m}{Maximum rubisco carboxylation rate at 25 Celsius for the
Medlyn et al. 2002 modified Arrhenius model}

\item{Vcmax_dS_m}{Entropy parameter of Vcmax (kJ mol-1) for the
Medlyn et al. 2002 modified Arrhenius model}

\item{Vcmax_Ea_m}{Activation energy of Vcmax (kJ mol-1) for the
Medlyn et al. 2002 modified Arrhenius model}

\item{Vcmax_Hd_m}{Deactivation energy of Vcmax (kJ mol-1) for the
Medlyn et al. 2002 modified Arrhenius model}

\item{Vcmax25_n}{Maximum rubisco carboxylation rate at 25 Celsius for the
rederived modified Arrhenius model}

\item{Vcmax_dS_n}{Entropy parameter of Vcmax (kJ mol-1) for the
rederived modified Arrhenius model}

\item{Vcmax_Ea_n}{Activation energy of Vcmax (kJ mol-1) for the
rederived modified Arrhenius model}

\item{Vcmax_Hd_n}{Deactivation energy of Vcmax (kJ mol-1) for the
rederived modified Arrhenius model}
}
\value{
Calculates photosynthetic parameters for the Farquhar et al. (1980)
model of photosynthesis. See ?t_functions for details on temperature scaling.

Farquhar GD, et al. 1980. A biochemical model of photosynthetic CO2 
assimilation in leaves of C3 species. Planta 149: 78-90.
}
\description{
Calculating photosynthetic parameters
}
