% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_photo.R
\name{calculate_photo}
\alias{calculate_photo}
\title{Title}
\usage{
calculate_photo(
  data_phys,
  data_env,
  varnames = list(GammaStar = "GammaStar", Jmax = "Jmax", Km = "Km", leaf_respiration =
    "leaf_respiration", root_respiration = "root_respiration", shoot_respiration =
    "shoot_respiration", total_respiration = "total_respiration", Ca = "Ca", Qin = "Qin",
    VPD = "VPD", Temp = "Temp", Vcmax = "Vcmax"),
  alpha,
  g0,
  g1,
  Oi,
  phi
)
}
\arguments{
\item{data_phys}{}

\item{data_env}{}

\item{varnames}{}

\item{alpha}{}

\item{g0}{}

\item{g1}{}

\item{Oi}{}

\item{phi}{}
}
\value{

}
\description{
Title
}
