% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_functions.R
\name{calculate_resp}
\alias{calculate_resp}
\title{Functions for calculating respiration}
\usage{
calculate_resp(
  leaf_Q10,
  root_Q10,
  shoot_Q10,
  leaf_k25,
  root_k25,
  shoot_k25,
  leaf_area,
  root_mass,
  shoot_mass,
  Temp
)
}
\arguments{
\item{leaf_Q10}{Leaf thermal sensitivity}

\item{root_Q10}{Root thermal sensitivity}

\item{shoot_Q10}{Shoot thermal sensitivity}

\item{leaf_k25}{Leaf respiration at 25 Celsius}

\item{root_k25}{Root respiration at 25 Celsius}

\item{shoot_k25}{Shoot respiration at 25 Celsius}

\item{leaf_area}{Leaf area in m2}

\item{root_mass}{Root mass in g}

\item{shoot_mass}{Shoot mass in g}

\item{Temp}{Temperature in Kelvin}
}
\value{
Calculates total respiration
}
\description{
Functions for calculating respiration
}
