% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t_functions.R
\name{Arrh_medlyn}
\alias{Arrh_medlyn}
\alias{Arrh_new}
\alias{Q10_funct}
\alias{Arrh_base}
\title{Temperature response functions}
\usage{
Arrh_medlyn(k25, delS, Ea, Hd, Temp)

Arrh_new(k25, delS, Ea, Hd, Temp)

Q10_funct(k25, Q10, Temp)

Arrh_base(k25, Ea, Temp)
}
\arguments{
\item{k25}{Rate of process at 25 Celsius}

\item{delS}{Entropy parameter in kJ/mol}

\item{Ea}{Activation enegy in kJ/mol}

\item{Hd}{Deactivation energy in kJ/mol}

\item{Temp}{Temperature in Kelvin}

\item{Q10}{Thermal sensitivity of process, representing X-fold change per
10 Celsius change in temperature}
}
\value{
Provides functions for modeling temperature responses of biological
processes

ADD REFERENCES HERE
}
\description{
Temperature response functions
}
